Orders ipv6 addresses before v4 in resolution.
When performing  mdns or dns resolution for a given hostname v4
addresses are always returned for the connection over v6 when an address
supports dual stacks.

This allows for ipv6 transitioning and testing in a dual stack
environment. It also follows common practice when picking the right
address to try first.

Not addressed by this commit is the ability to fall back to other
resolved addresses should the first address fail. This feature is
addressed by other todo's in the project.