Stop building wheels for python 3.9/3.10
When esphome 2023.11.0 is released, the containers will be
python 3.11 based so we will no longer need wheels for old
versions.

Note that this does not increase the library minimum python
version and wheels can still be built locally for older python
versions.

This change is only to decrease the time it takes to do a release
and save space on PyPI