Refactor frame helper to allow sending multiple packets at once
Since we we TCP_NODELAY it means all packets get written to
the wire as soon as the buffer has space which means they
get written almost right away in most cases. When we know
we need to send multiple packets there was no way to send
them in one write which increased protocol latancy. To
fix this write_packet is now write_packets which will do
the write to the transport in a single go which increases
the change the ESP will not have to transfer data from
the stack multiple times to read incoming data