Refactor raw Bluetooth advs subscriptions
Currently we only pass a list of BluetoothLERawAdvertisement back
to the callback, which makes it difficult to change without a breaking
change later when we need to add more information about the
adv (such as BLE version, etc). Instead we now pass the
BluetoothLERawAdvertisementsResponse object. This means we
leak the abstraction a bit more, but considering how tightly
coupled this all is, that seems like a good trade off to
avoid a future breaking change later now that we will use
this in bleak-esphome