Make CameraImageResponse and SubscribeLogsResponse zero_copy
Currently we have manually written methods for these messages in
esphome. We want to remove the manual code and use these objects
directly since we generate code for them. We had manual methods
to avoid the copy but since we can mark them zero copy and generate
different code, we avoid the duplicate code